+++ Modules/Intent.Modules.IaC.Terraform/Templates/Subscriptions/AzureEventGridTf/AzureEventGridTfTemplatePartial.cs
@@ -13,6 +13,7 @@ using Intent.Modules.Common.CSharp.TypeResolvers;
 using Intent.Modules.Common.Templates;
 using Intent.Modules.Common.Types.Api;
 using Intent.Modules.Integration.IaC.Shared;
+using Intent.Modules.Integration.IaC.Shared.AzureEventGrid;
 using Intent.RoslynWeaver.Attributes;
 using Intent.Templates;
 
@@ -57,8 +58,8 @@ namespace Intent.Modules.IaC.Terraform.Templates.Subscriptions.AzureEventGridTf
             foreach (var topicSubscription in topicSubscriptions)
             {
                 var subscription = topicSubscription.First();
-                var topicNameRef = subscription.SubscriptionItem.TopicName.ToSnakeCase();
-                builder.AddData(Terraform.azurerm_eventgrid_topic.type, topicNameRef, data =>
+                var message = subscription.SubscriptionItem;
+                builder.AddData(Terraform.azurerm_eventgrid_topic.type, Terraform.azurerm_eventgrid_topic.topic(message).refname, data =>
                 {
                     data.AddSetting("name", subscription.SubscriptionItem.TopicName);
                     data.AddRawSetting("resource_group_name", "var.resource_group_name");
@@ -66,8 +67,8 @@ namespace Intent.Modules.IaC.Terraform.Templates.Subscriptions.AzureEventGridTf
 
                 builder.AddResource(Terraform.azurerm_eventgrid_event_subscription.type, $"{_sanitizedApplicationName.ToSnakeCase()}", resource =>
                 {
-                    resource.AddSetting("name", $"{subscription.SubscriptionItem.TopicName.ToKebabCase()}-sub");
-                    resource.AddRawSetting("scope", $"data.{Terraform.azurerm_eventgrid_topic.type}.{topicNameRef}.id");
+                    resource.AddSetting("name", AzureHelper.EnsureValidLength(Model.Name, 64));
+                    resource.AddRawSetting("scope", $"data.{Terraform.azurerm_eventgrid_topic.topic(message).id}");
                     resource.AddSetting("event_delivery_schema", "CloudEventSchemaV1_0");
                     resource.AddBlock("azure_function_endpoint", b =>
                     {