+++ Modules/Intent.Modules.Integration.IaC.Shared.AzureEventGrid/IntegrationManager.cs
@@ -47,7 +47,7 @@ internal class IntegrationManager
             .Where(app => app.Modules.Any(x => x.ModuleId == azureEventGridModule))
             .SelectMany(app => executionContext.MetadataManager
                 .GetExplicitlyPublishedMessageModels(app.Id)
-                .Select(message => new MessageInfo(app.Id, message, null)))
+                .Select(message => new MessageInfo(app.Id, app.Name, message, null)))
             .Distinct()
             .ToList();
         _subscribedMessages = applications
@@ -61,7 +61,7 @@ internal class IntegrationManager
                         Message = sub.Element.AsMessageModel(),
                         Handler = handler
                     }))
-                .Select(sub => new MessageInfo(app.Id, sub.Message, sub.Handler)))
+                .Select(sub => new MessageInfo(app.Id, app.Name, sub.Message, sub.Handler)))
             .Distinct()
             .ToList();
         
@@ -69,7 +69,7 @@ internal class IntegrationManager
             .Where(app => app.Modules.Any(x => x.ModuleId == azureEventGridModule))
             .SelectMany(app => executionContext.MetadataManager
                 .GetExplicitlySentIntegrationCommandModels(app.Id)
-                .Select(command => new CommandInfo(app.Id, command, null)))
+                .Select(command => new CommandInfo(app.Id, app.Name, command, null)))
             .Distinct()
             .ToList();
         _receivedCommands = applications
@@ -83,7 +83,7 @@ internal class IntegrationManager
                         Command = sub.Element.AsIntegrationCommandModel(),
                         Handler = handler
                     }))
-                .Select(sub => new CommandInfo(app.Id, sub.Command, sub.Handler)))
+                .Select(sub => new CommandInfo(app.Id, app.Name, sub.Command, sub.Handler)))
             .Distinct()
             .ToList();
     }
@@ -92,7 +92,7 @@ internal class IntegrationManager
     {
         var messages = _publishedMessages
             .Where(p => p.ApplicationId == applicationId)
-            .Select(s => new AzureEventGridMessage(s.ApplicationId, s.Message, AzureEventGridMethodType.Publish))
+            .Select(s => new AzureEventGridMessage(s.ApplicationId, s.ApplicationName, s.Message, AzureEventGridMethodType.Publish))
             .ToList();
         return messages;
     }
@@ -101,7 +101,7 @@ internal class IntegrationManager
     {
         var messages = _subscribedMessages
             .Where(p => p.ApplicationId == applicationId)
-            .Select(s => new AzureEventGridMessage(s.ApplicationId, s.Message, AzureEventGridMethodType.Subscribe))
+            .Select(s => new AzureEventGridMessage(s.ApplicationId, s.ApplicationName, s.Message, AzureEventGridMethodType.Subscribe))
             .ToList();
         return messages;
     }
@@ -121,7 +121,7 @@ internal class IntegrationManager
             .Where(message => message.ApplicationId == applicationId)
             .Select(message => new Subscription<AzureEventGridMessage>(
                 message.EventHandlerModel!,
-                new AzureEventGridMessage(applicationId, message.Message, AzureEventGridMethodType.Subscribe)))
+                new AzureEventGridMessage(applicationId, message.ApplicationName, message.Message, AzureEventGridMethodType.Subscribe)))
         );
 
         return results;
@@ -129,6 +129,6 @@ internal class IntegrationManager
 
     public record Subscription<TSubscriptionItem>(IntegrationEventHandlerModel EventHandlerModel, TSubscriptionItem SubscriptionItem);
     
-    private record MessageInfo(string ApplicationId, MessageModel Message, IntegrationEventHandlerModel? EventHandlerModel);
-    private record CommandInfo(string ApplicationId, IntegrationCommandModel Command, IntegrationEventHandlerModel? EventHandlerModel);
+    private record MessageInfo(string ApplicationId, string ApplicationName, MessageModel Message, IntegrationEventHandlerModel? EventHandlerModel);
+    private record CommandInfo(string ApplicationId, string ApplicationName, IntegrationCommandModel Command, IntegrationEventHandlerModel? EventHandlerModel);
 }