+++ Tests/AspNetCore.AzureServiceBus.GroupA/AspNetCore.AzureServiceBus.GroupA.Application/CreateClient/CreateClientCommandHandler.cs
@@ -1,0 +1,31 @@
+using AspNetCore.AzureServiceBus.GroupA.Application.Common.Eventing;
+using AspNetCore.AzureServiceBus.GroupA.Eventing.Messages;
+using Intent.RoslynWeaver.Attributes;
+using MediatR;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Application.MediatR.CommandHandler", Version = "2.0")]
+
+namespace AspNetCore.AzureServiceBus.GroupA.Application.CreateClient
+{
+    [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
+    public class CreateClientCommandHandler : IRequestHandler<CreateClientCommand>
+    {
+        private readonly IEventBus _eventBus;
+
+        [IntentManaged(Mode.Merge)]
+        public CreateClientCommandHandler(IEventBus eventBus)
+        {
+            _eventBus = eventBus;
+        }
+
+        [IntentManaged(Mode.Fully, Body = Mode.Fully)]
+        public async Task Handle(CreateClientCommand request, CancellationToken cancellationToken)
+        {
+            _eventBus.Publish(new ClientCreatedEvent
+            {
+                Name = request.Name
+            });
+        }
+    }
+}