+++ Tests/AspNetCore.AzureServiceBus.GroupA/AspNetCore.AzureServiceBus.GroupA.Application/DependencyInjection.cs
@@ -1,0 +1,29 @@
+using System.Reflection;
+using AspNetCore.AzureServiceBus.GroupA.Application.Common.Behaviours;
+using Intent.RoslynWeaver.Attributes;
+using MediatR;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Application.DependencyInjection.DependencyInjection", Version = "1.0")]
+
+namespace AspNetCore.AzureServiceBus.GroupA.Application
+{
+    public static class DependencyInjection
+    {
+        public static IServiceCollection AddApplication(this IServiceCollection services, IConfiguration configuration)
+        {
+            services.AddMediatR(cfg =>
+            {
+                cfg.RegisterServicesFromAssembly(Assembly.GetExecutingAssembly());
+                cfg.AddOpenBehavior(typeof(UnhandledExceptionBehaviour<,>));
+                cfg.AddOpenBehavior(typeof(PerformanceBehaviour<,>));
+                cfg.AddOpenBehavior(typeof(AuthorizationBehaviour<,>));
+                cfg.AddOpenBehavior(typeof(EventBusPublishBehaviour<,>));
+                cfg.AddOpenBehavior(typeof(UnitOfWorkBehaviour<,>));
+            });
+            return services;
+        }
+    }
+}