+++ Tests/AspNetCore.AzureServiceBus.GroupA/AspNetCore.AzureServiceBus.GroupA.Infrastructure/Configuration/AzureServiceBusConfiguration.cs
@@ -1,0 +1,31 @@
+using AspNetCore.AzureServiceBus.GroupA.Application.Common.Eventing;
+using AspNetCore.AzureServiceBus.GroupA.Eventing.Messages;
+using AspNetCore.AzureServiceBus.GroupA.Infrastructure.Eventing;
+using Azure.Messaging.ServiceBus;
+using Intent.RoslynWeaver.Attributes;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Eventing.AzureServiceBus.AzureServiceBusConfiguration", Version = "1.0")]
+
+namespace AspNetCore.AzureServiceBus.GroupA.Infrastructure.Configuration
+{
+    public static class AzureServiceBusConfiguration
+    {
+        public static IServiceCollection ConfigureAzureServiceBus(
+            this IServiceCollection services,
+            IConfiguration configuration)
+        {
+            services.AddSingleton<ServiceBusClient>(sp => new ServiceBusClient(configuration["AzureServiceBus:ConnectionString"]));
+            services.AddScoped<IEventBus, AzureServiceBusEventBus>();
+            services.AddSingleton<AzureServiceBusMessageDispatcher>();
+            services.AddSingleton<IAzureServiceBusMessageDispatcher, AzureServiceBusMessageDispatcher>();
+            services.Configure<AzureServiceBusPublisherOptions>(options =>
+            {
+                options.Add<ClientCreatedEvent>(configuration["AzureServiceBus:ClientCreated"]!);
+            });
+            return services;
+        }
+    }
+}