+++ Tests/Infrastructure.Terraform/01-applications/azure-functions-azure-service-bus-group-b.tf
@@ -52,10 +52,10 @@ resource "azurerm_windows_function_app" "azure-functions-azure-service-bus-group
     "FUNCTIONS_WORKER_RUNTIME"                  = "dotnet-isolated"
     "AzureServiceBus:ConnectionString"          = azurerm_servicebus_namespace.service_bus.default_primary_connection_string
     "AzureServiceBus:SpecificTopic"             = azurerm_servicebus_topic.specific_topic.name
-    "AzureServiceBus:SpecificQueue"             = azurerm_servicebus_topic.specific_queue.name
+    "AzureServiceBus:SpecificQueue"             = azurerm_servicebus_queue.specific_queue.name
     "AzureServiceBus:ClientCreated"             = azurerm_servicebus_topic.client_created.name
-    "AzureServiceBus:ClientCreatedSubscription" = azurerm_servicebus_subscription.client_created.name
-    "AzureServiceBus:CreateOrg"                 = azurerm_servicebus_topic.create_org.name
+    "AzureServiceBus:ClientCreatedSubscription" = azurerm_servicebus_subscription.azure_functions_azure_service_bus_group_b_client_created.name
+    "AzureServiceBus:CreateOrg"                 = azurerm_servicebus_queue.create_org.name
   }
 }