+++ Modules/Intent.Modules.AspNetCore.IdentityService/release-notes.md
@@ -1,3 +1,5 @@
 ### Version 1.0.0
 
+Fixed: If uninstalled and re-installed will no longer create duplicates and will only install in first service package.
+
 Initial Release